+++ SIL.Media/FFmpegRunner.cs
@@ -1,13 +1,16 @@
 using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
+using System.Linq;
+using System.Text.RegularExpressions;
 using JetBrains.Annotations;
 using SIL.CommandLineProcessing;
 using SIL.IO;
 using SIL.PlatformUtilities;
 using SIL.Progress;
+using static System.Environment;
+using static System.IO.Path;
 using static System.String;
+using Version = System.Version;
 
 namespace SIL.Media
 {
@@ -18,15 +21,41 @@ namespace SIL.Media
 	/// maybe figure out how to use that to provide progress reporting, but it's not clear
 	/// that it would be worth it.
 	///</summary>
-	public class FFmpegRunner
+	public static class FFmpegRunner
 	{
+		internal const string kLinuxBinFolder = "/usr/bin";
+		internal const string kFFmpeg = "ffmpeg";
 		private const string kFFmpegExe = "ffmpeg.exe";
-		private const string mp3LameCodecArg = "-acodec libmp3lame";
+		private const string kMp3LameCodecArg = "-acodec libmp3lame";
 
 		/// <summary>
-		/// If your app knows where FFmpeg lives, you can tell us before making any calls.
+		/// If your app knows where FFmpeg lives, set this before making any calls.
+		/// Unless the exe is on the system path, this is the full path to the executable,
+		/// including the executable file itself.
 		/// </summary>
 		public static string FFmpegLocation;
+		private static bool s_locationSetByClient = true;
+
+		/// <summary>
+		/// If your app has a known minimum version of FFMpeg tools that it will work with, you can
+		/// set this to prevent this library from attempting to use a version that is not going to
+		/// meet your requirements. This will be ignored if you set FFmpegLocation, if the ffmpeg
+		/// installation is based on a Linux package dependency, or if FFmpeg is colocated with the
+		/// applications, since it seems safe to assume that you would not specify or install a
+		/// version that does not satisfy your needs. Note that this also applies to FFprobe (used
+		/// in MediaInfo).
+		/// </summary>
+		public static Version MinimumVersion
+		{
+			get => s_minimumVersion;
+			set
+			{
+				s_minimumVersion = value;
+				if (!s_locationSetByClient)
+					FFmpegLocation = null;
+			}
+		}
+		private static Version s_minimumVersion;
 
 		/// <summary>
 		/// Find the path to FFmpeg, and remember it (some apps (like SayMore) call FFmpeg a lot)
@@ -34,76 +63,113 @@ namespace SIL.Media
 		/// <returns></returns>
 		internal static string LocateAndRememberFFmpeg()
 		{
-			if (null != FFmpegLocation) //NO! string.empty means we looked and didn't find: string.IsNullOrEmpty(s_ffmpegLocation))
+			// Do not change this to IsNullOrEmpty(FFmpegLocation) because Empty means we already
+			// looked and didn't find it.
+			if (null != FFmpegLocation)
 				return FFmpegLocation;
-			FFmpegLocation = LocateFFmpeg() ?? Empty;
+
+			s_locationSetByClient = false;
+
+			FFmpegLocation = LocateFFmpeg();
 			return FFmpegLocation;
+
+			// This returns the exe path if found; otherwise Empty.
+			static string LocateFFmpeg()
+			{
+				if (Platform.IsLinux)
+				{
+					// On Linux, we can assume the package has included the needed dependency.
+					var convExePath = Combine(kLinuxBinFolder, kFFmpeg);
+					if (File.Exists(convExePath))
+						return convExePath;
+					// Try avconv, the new name of ffmpeg on Linux
+					convExePath = Combine(kLinuxBinFolder, "avconv");
+					return File.Exists(convExePath) ? convExePath : Empty;
+				}
+
+				// On Windows FFmpeg will typically be distributed with the SIL software that
+				// accompanies the SIL.Media DLL.
+				var withApplicationDirectory = GetPathToBundledFFmpeg();
+				if (withApplicationDirectory != null && File.Exists(withApplicationDirectory))
+					return withApplicationDirectory;
+
+				// Failing that, if a program wants to use this library and work with a version of
+				// it that the user downloaded or compiled locally, this logic tries to find it.
+				var fromChoco = GetFFmpegFolderFromChocoInstall(kFFmpegExe);
+				if (fromChoco != null)
+				{
+					var pathToFFmpeg = Combine(fromChoco, kFFmpegExe);
+					return pathToFFmpeg;
+				}
+
+				// Try to just run ffmpeg from the path, if it works then we can use that directly.
+				if (MeetsMinimumVersionRequirement(kFFmpeg))
+					return kFFmpeg;
+
+				// REVIEW: I just followed the instructions in the current version of Audacity for
+				// installing FFmpeg for Audacity and the result is a folder that does not contain
+				// ffmpeg.exe. This maybe used to work, but I don't think we'll ever find ffmpeg this
+				// way now.
+				// https://support.audacityteam.org/basics/installing-ffmpeg
+				return new[] {
+						GetFolderPath(SpecialFolder.ProgramFiles),
+						GetFolderPath(SpecialFolder.ProgramFilesX86) }
+					.Select(path => Combine(path, "FFmpeg for Audacity", kFFmpegExe))
+					.FirstOrDefault(exePath => File.Exists(exePath) &&
+						MeetsMinimumVersionRequirement(exePath)) ?? Empty;
+			}
 		}
 
-		/// <summary>
-		/// FFmpeg will typically be distributed with SIL software on Windows or automatically
-		/// installed via package dependencies on other platforms, but if something wants to
-		/// use this library and work with a version of it that the user downloaded or compiled
-		/// locally, this tries to find where they put it.
-		/// </summary>
-		/// <returns>the path, if found, else null</returns>
-		private static string LocateFFmpeg()
+		internal static bool MeetsMinimumVersionRequirement(string exe)
 		{
-			if (Platform.IsLinux)
+			try
 			{
-				//on linux, we can safely assume the package has included the needed dependency
-				if (File.Exists("/usr/bin/ffmpeg"))
-					return "/usr/bin/ffmpeg";
-				if (File.Exists("/usr/bin/avconv"))
-					return "/usr/bin/avconv"; // the new name of ffmpeg on Linux
+				var version = new Regex(GetFileNameWithoutExtension(exe).ToLowerInvariant() +
+					@" version (?<version>\d+\.\d+(\.\d+)?)");
+				var results = CommandLineRunner.Run(exe, "-version", ".", 5, new NullProgress());
+				var match = version.Match(results.StandardOutput);
+				if (!match.Success)
+					return false;
+				if (MinimumVersion == null)
+					return true;
+				var actualVersion = Version.Parse(match.Groups["version"].Value);
+				actualVersion = new Version(actualVersion.Major, actualVersion.Minor,
+					actualVersion.Build >= 0 ? actualVersion.Build : 0,
+					actualVersion.Revision >= 0 ? actualVersion.Revision : 0);
+				return actualVersion >= MinimumVersion;
+			}
+			catch
+			{
+				return false;
+			}
+		}
 
+		internal static string GetFFmpegFolderFromChocoInstall(string exeNeeded)
+		{
+			try
+			{
+				var programData = GetFolderPath(SpecialFolder
+					.CommonApplicationData);
+
+				var folder = Combine(programData, "chocolatey", "lib", kFFmpeg, "tools",
+					kFFmpeg, "bin");
+				var pathToExe = Combine(folder, exeNeeded);
+				if (!File.Exists(pathToExe)|| !MeetsMinimumVersionRequirement(pathToExe))
+					folder = null;
+				return folder;
+			}
+			catch (Exception e)
+			{
+				Console.WriteLine(e);
 				return null;
 			}
-
-			string withApplicationDirectory = GetPathToBundledFFmpeg();
-
-			if (withApplicationDirectory != null && File.Exists(withApplicationDirectory))
-				return withApplicationDirectory;
-
-			var fromChoco = MediaInfo.GetFFmpegFolderFromChocoInstall(kFFmpegExe);
-			if (fromChoco != null)
-				return Path.Combine(fromChoco, kFFmpegExe);
-
-			var progFileDirs = new List<string> {
-				Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86),
-				Environment.GetFolderPath(Environment.SpecialFolder.ProgramFiles)
-			};
-
-			/* We DON'T SUPPORT THIS ONE (it lacks some information on the output, at least as of
-			 * July 2010)
-			 //from http://www.arachneweb.co.uk/software/windows/avchdview/ffmpeg.html
-			foreach (var path in progFileDirs)
-			{
-				var exePath = (Path.Combine(path, "ffmpeg/win32-static/bin/ffmpeg.exe"));
-				if(File.Exists(exePath))
-					return exePath;
-			}
-			 */
-
-			// REVIEW: I just followed the instructions in the current version of Audacity for
-			// installing FFmpeg for Audacity and the result is a folder that does not contain
-			// ffmpeg.exe. This maybe used to work, but I don't think we'll ever find ffmpeg this
-			// way now.
-			//http://manual.audacityteam.org/index.php?title=FAQ:Installation_and_Plug-Ins#installffmpeg
-			foreach (var path in progFileDirs)
-			{
-				var exePath = (Path.Combine(path, "FFmpeg for Audacity", kFFmpegExe));
-				if (File.Exists(exePath))
-					return exePath;
-			}
-			return null;
 		}
 
 		private static string GetPathToBundledFFmpeg()
 		{
 			try
 			{
-				return FileLocationUtilities.GetFileDistributedWithApplication("ffmpeg", kFFmpegExe);
+				return FileLocationUtilities.GetFileDistributedWithApplication(kFFmpeg, kFFmpegExe);
 			}
 			catch (Exception)
 			{
@@ -114,42 +180,10 @@ namespace SIL.Media
 		///<summary>
 		/// Returns false if it can't find ffmpeg
 		///</summary>
-		public static bool HaveNecessaryComponents => LocateFFmpeg() != null;
+		public static bool HaveNecessaryComponents => LocateAndRememberFFmpeg() != Empty;
 
 		private static ExecutionResult NoFFmpeg =>
 			new ExecutionResult { StandardError = "Could not locate FFmpeg" };
-
-		///<summary>
-		/// Returns false if it can't find ffmpeg
-		///</summary>
-		private static bool HaveValidFFmpegOnPath
-		{
-			get
-			{
-				if (Platform.IsWindows)
-				{
-					if (LocateFFmpeg() != null)
-						return true;
-				}
-
-				//see if there is one on the %path%
-
-				var p = new Process();
-				p.StartInfo.FileName = "ffmpeg";
-				p.StartInfo.RedirectStandardError = true;
-				p.StartInfo.UseShellExecute = false;
-				p.StartInfo.CreateNoWindow = true;
-				try
-				{
-					p.Start();
-				}
-				catch (Exception)
-				{
-					return false;
-				}
-				return true;
-			}
-		}
 
 		/// <summary>
 		/// Extracts the audio from a video. Note, it will fail if the file exists, so the client
@@ -162,13 +196,13 @@ namespace SIL.Media
 		/// <returns>log of the run</returns>
 		public static ExecutionResult ExtractMp3Audio(string inputPath, string outputPath, int channels, IProgress progress)
 		{
-			if (LocateFFmpeg() == null)
+			if (LocateAndRememberFFmpeg() == null)
 				return NoFFmpeg;
 
-			var arguments = $"-i \"{inputPath}\" -vn {mp3LameCodecArg} -ac {channels} \"{outputPath}\"";
+			var arguments = $"-i \"{inputPath}\" -vn {kMp3LameCodecArg} -ac {channels} \"{outputPath}\"";
 			var result = RunFFmpeg(arguments, progress);
 
-			//hide a meaningless error produced by some versions of liblame
+			// Hide a meaningless error produced by some versions of liblame
 			if (result.StandardError.Contains("lame: output buffer too small")
 				&& File.Exists(outputPath))
 			{
@@ -197,13 +231,13 @@ namespace SIL.Media
 		/// <returns>log of the run</returns>
 		public static ExecutionResult ExtractOggAudio(string inputPath, string outputPath, int channels, IProgress progress)
 		{
-			if (LocateFFmpeg() == null)
+			if (LocateAndRememberFFmpeg() == null)
 				return NoFFmpeg;
 
 			var arguments = $"-i \"{inputPath}\" -vn -acodec vorbis -ac {channels} \"{outputPath}\"";
 			var result = RunFFmpeg(arguments, progress);
 
-			//hide a meaningless error produced by some versions of liblame
+			// Hide a meaningless error produced by some versions of liblame
 			if (result.StandardError.Contains("lame: output buffer too small")
 				&& File.Exists(outputPath))
 			{
@@ -278,7 +312,7 @@ namespace SIL.Media
 		private static ExecutionResult ExtractAudio(string inputPath, string outputPath,
 			string audioCodec, int sampleRate, int channels, IProgress progress)
 		{
-			if (LocateFFmpeg() == null)
+			if (LocateAndRememberFFmpeg() == null)
 				return NoFFmpeg;
 
 			var sampleRateArg = "";
@@ -294,7 +328,7 @@ namespace SIL.Media
 
 			var result = RunFFmpeg(arguments, progress);
 
-			//hide a meaningless error produced by some versions of liblame
+			// Hide a meaningless error produced by some versions of liblame
 			if (result.StandardError.Contains("lame: output buffer too small")
 				&& File.Exists(outputPath))
 			{
@@ -321,14 +355,14 @@ namespace SIL.Media
 		public static ExecutionResult ChangeNumberOfAudioChannels(string inputPath,
 			string outputPath, int channels, IProgress progress)
 		{
-			if (LocateFFmpeg() == null)
+			if (LocateAndRememberFFmpeg() == null)
 				return NoFFmpeg;
 
 			var arguments = $"-i \"{inputPath}\" -vn -ac {channels} \"{outputPath}\"";
 
 			var result = RunFFmpeg(arguments, progress);
 
-			//hide a meaningless error produced by some versions of liblame
+			// Hide a meaningless error produced by some versions of liblame
 			if (result.StandardError.Contains("lame: output buffer too small") && File.Exists(outputPath))
 			{
 				var doctoredResult = new ExecutionResult
@@ -357,11 +391,11 @@ namespace SIL.Media
 			if (IsNullOrEmpty(LocateAndRememberFFmpeg()))
 				return NoFFmpeg;
 
-			var arguments = $"-i \"{inputPath}\" {mp3LameCodecArg} -ac 1 -ar 8000 \"{outputPath}\"";
+			var arguments = $"-i \"{inputPath}\" {kMp3LameCodecArg} -ac 1 -ar 8000 \"{outputPath}\"";
 
 			var result = RunFFmpeg(arguments, progress);
 
-			//hide a meaningless error produced by some versions of liblame
+			// Hide a meaningless error produced by some versions of liblame
 			if (result.StandardError.Contains("lame: output buffer too small")
 				&& File.Exists(outputPath))
 			{
@@ -397,14 +431,14 @@ namespace SIL.Media
 
 			// isn't working: var arguments = "-i \"" + inputPath + "\" -vcodec mpeg4 -s 160x120 -b 800  -acodec libmp3lame -ar 22050 -ab 32k -ac 1 \"" + outputPath + "\"";
 			var arguments = $"-i \"{inputPath}\" -vcodec mpeg4 -s 160x120 -b 800 " +
-				$"{mp3LameCodecArg} -ar 22050 -ab 32k -ac 1 ";
+				$"{kMp3LameCodecArg} -ar 22050 -ab 32k -ac 1 ";
 			if (maxSeconds > 0)
 				arguments += $" -t {maxSeconds} ";
 			arguments += $"\"{outputPath}\"";
 
 			var result = RunFFmpeg(arguments, progress);
 
-			//hide a meaningless error produced by some versions of liblame
+			// Hide a meaningless error produced by some versions of liblame
 			if (result.StandardError.Contains("lame: output buffer too small")
 				&& File.Exists(outputPath))
 			{
@@ -448,11 +482,11 @@ namespace SIL.Media
 
 		private static ExecutionResult RunFFmpeg(string arguments, IProgress progress)
 		{
-			progress.WriteMessage("ffmpeg " + arguments);
+			progress.WriteMessage($"{GetFileNameWithoutExtension(LocateAndRememberFFmpeg())} {arguments}");
 
 			const int timeout = 600; // 60 * 10 = 10 minutes
 			var result = CommandLineRunner.Run(LocateAndRememberFFmpeg(), arguments,
-				Environment.CurrentDirectory, timeout, progress);
+				CurrentDirectory, timeout, progress);
 
 			progress.WriteVerbose(result.StandardOutput);